@@ -2,6 +2,12 @@ import {defineConfig} from '@rsbuild/core';
 import {pluginSvelte} from '@rsbuild/plugin-svelte';
 import * as process from "node:process";
 import {pluginCssMinimizer} from "@rsbuild/plugin-css-minimizer";
+import { convert } from 'tsconfig-to-swcconfig';
+import { resolve } from 'path';
+
+const swcConfig = convert("tsconfig.json"); // This will look for tsconfig.json in the current directory
+swcConfig.env = null;
+swcConfig.jsc.baseUrl = resolve(__dirname, 'src')
 
 export default defineConfig({
     environments: {
@@ -25,6 +31,8 @@ export default defineConfig({
     dev: {
         hmr: false,
         liveReload: true,
+    },
+    tools: {
+        swc: swcConfig,
     }
-    //TODO npx run stage fails with root: './src'. This makes no sense
 });
