@@ -1,9 +1,9 @@
 <script lang="ts">
-  import { appState } from "ts/state.ts";
+  import { appState } from "ts/state.svelte.ts";
   import { API } from "./ts/api.ts";
   // noinspection TypeScriptCheckImport
   //import QR from "svelte-qr";
-  import { formatBalance, minBalanceName, qrString } from "./services";
+  import { formatBalance, minBalanceName, qrString } from "./ts/services.svelte.ts";
   import type { PaymentResponse } from "./types/backend";
 
   export let total: number;
@@ -37,7 +37,7 @@
   <div class="container">
     <div class="p-2">
       <select bind:value={selected}>
-        {#each Object.entries(appState.$state.config.supportedCurrencies) as [key, value]}
+        {#each Object.entries(appState.config.supportedCurrencies) as [key, value]}
           {#if value.isCrypto}
             <option value={key}>
               {value.name}
